+++ src/Form/Extension/GraphQLExtensionType.php
@@ -42,7 +42,7 @@ class GraphQLExtensionType extends AbstractTypeExtension
     /**
      * {@inheritdoc}
      */
-    public static function getExtendedTypes()
+    public static function getExtendedTypes(): iterable
     {
         return [FormType::class];
     }