+++ src/YnloGraphQLBundle.php
@@ -11,7 +11,6 @@
 namespace Ynlo\GraphQLBundle;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Ynlo\GraphQLBundle\Component\TaggedServices\TaggedServicesCompilerPass;
 use Ynlo\GraphQLBundle\DependencyInjection\Compiler\ControllerPass;
@@ -23,8 +22,6 @@ use Ynlo\GraphQLBundle\Filter\FilterInterface;
 use Ynlo\GraphQLBundle\Filter\FilterResolverInterface;
 use Ynlo\GraphQLBundle\Form\Input\InputFieldTypeGuesser;
 use Ynlo\GraphQLBundle\SearchBy\SearchByInterface;
-use Ynlo\GraphQLBundle\Subscription\Publisher;
-use Ynlo\GraphQLBundle\Subscription\SubscriptionAwareInterface;
 use Ynlo\GraphQLBundle\Type\Loader\TypeAutoLoader;
 use Ynlo\GraphQLBundle\Type\Registry\TypeRegistry;
 use Ynlo\GraphQLBundle\Util\IDEncoder;